Move rendering unit tests in dedicated module. Add XFAIL flag to rendering tests.
